# docs/openapi.yaml
openapi: "3.1.0"
...
info:
  title: Chronos Ledger API
  version: "0.10.0" # x-release-please-version
  description: |
    Organization Schedule & Attendance Management REST API.

    **Authentication**: All endpoints except `/config` and `/auth/login`
    require a Bearer JWT in the `Authorization` header. An API key in
    `X-API-Key` is accepted in place of one anywhere, which is how a lobby
    terminal with no one to sign in reaches `/guest/register-checkin` and
    `/guest/directory`. A key can be narrowed to part of the API and given
    an end date: see the `ApiKeyAuth` security scheme.

    **Roles**: `SUPER_ADMIN` > `UNIT_ADMIN` > `STAFF` > `MEMBER`.
    Role requirements are noted on each endpoint.
  contact:
    name: Chronos Ledger Contributors
    email: admin@org.internal
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost
    description: Local development (Docker Compose)
  - url: http://{server_ip}
    description: Organization intranet
    variables:
      server_ip:
        default: "192.168.1.100"

# ─────────────────────────────────────────────
# Reusable components
# ─────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        A machine credential bound to a service-account user. Accepted anywhere
        a Bearer token is, and acts as the bound user, so roles and unit scoping
        apply unchanged. This is how a lobby kiosk authenticates.

        A key can also be narrowed and given an end date, which a Bearer token
        cannot be. `scopes` is a comma separated list of `<area>:read` and
        `<area>:write`, where the area is the path segment after `/api/v1`:
        `schedule`, `resources`, `attendance`, `users`, `ingestion`, `guest`,
        `sync`, `config`, `auth`, `api-keys`. `GET`, `HEAD` and `OPTIONS` are
        reads and everything else is a write, so `schedule:read` passes
        `GET /schedule/slots` and is refused `POST /schedule/slots` with a 403
        naming `schedule:write`. The single scope `*` is every area, which is
        what a key issued before scopes existed holds and what a key created
        without naming any gets.

        `expires_at` is checked on the request, not by a sweep, so a key stops
        working the moment it runs out and answers 401 `API key expired`. A
        null expiry never runs out.

        The first-login gate does not apply to a key. It holds an admin
        account until the password it was handed is changed, and a key never
        uses that password, so a service account at an admin role works
        through its key without anyone signing in as it.
  schemas:
...
    InstitutionalRole:
      type: string
      enum: [SUPER_ADMIN, UNIT_ADMIN, STAFF, MEMBER]
      description: Role assigned to a system user.
    AccessReadiness:
      type: string
      enum: [OPEN_AD_HOC, BUSY, CRITICAL_DO_NOT_DISTURB, VERY_FREE]
      description: Staff availability / occupancy index visible to members.
...
    UserCreate:
      type: object
      required: [id, full_name, email_address, password, role_type]
      properties:
        id:
          type: string
          example: FAC042
        full_name:
          type: string
          example: Dr. Rajan Mehta
        email_address:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 12
          description: >
            Held to the same PASSWORD_MIN_LENGTH floor as a password the user
            chooses for themselves.
        role_type:
          $ref: "#/components/schemas/InstitutionalRole"
        unit_code:
          type: string
          nullable: true
          example: CSE
        assigned_base_station:
          type: string
          nullable: true
          example: Front Desk
        reporting_line_manager:
          type: string
          nullable: true
          example: HOD001
    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
          nullable: true
        email_address:
          type: string
          format: email
          nullable: true
        unit_code:
          type: string
          nullable: true
        assigned_base_station:
          type: string
          nullable: true
        reporting_line_manager:
          type: string
          nullable: true
...
    PasswordResetResponse:
      type: object
      properties:
        user_id:
          type: string
        initial_password:
          type: string
          description: >
            Plain text, returned only here. Never stored and never logged.
      required: [user_id, initial_password]
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/AccessReadiness"
    UserResponse:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        email_address:
          type: string
        role_type:
          $ref: "#/components/schemas/InstitutionalRole"
        unit_code:
          type: string
          nullable: true
        assigned_base_station:
          type: string
          nullable: true
        current_occupancy_index:
          $ref: "#/components/schemas/AccessReadiness"
        reporting_line_manager:
          type: string
          nullable: true
        deactivated_at:
          type: string
          format: date-time
          nullable: true
          description: When the account was deactivated, or null while it is active.

    # ── API keys ─────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: Not authorized.
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but insufficient role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ─────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────
paths:
...
  /api/v1/users/:
    get:
      tags: [Users]
      summary: List all users
      description: Requires `SUPER_ADMIN` or `UNIT_ADMIN`.
      parameters:
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/InstitutionalRole"
          description: Filter by role.
        - name: unit
          in: query
          schema:
            type: string
          description: Filter by unit code.
        - name: include_deactivated
          in: query
          schema:
            type: boolean
            default: false
          description: Include deactivated accounts, which are left out otherwise.
      responses:
        "200":
          description: User list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Users]
      summary: Create a new user
      description: Requires `SUPER_ADMIN`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The manager named in `reporting_line_manager` does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate user ID or email.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: >
            A malformed body, or a `reporting_line_manager` that names the
            user, someone who already reports to them, or someone who has been
            deactivated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/users/{user_id}:
    get:
      tags: [Users]
      summary: Get a specific user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Users]
      summary: Update a user
      description: >
        Requires `SUPER_ADMIN` or `UNIT_ADMIN`, and a unit admin only within
        their own unit. Only a super-admin can modify an account that is itself
        an admin. A user cannot update their own record here.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The user, or the manager named in `reporting_line_manager`, does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The email address is already registered to another account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: >
            A malformed body, or a `reporting_line_manager` that names the
            user, someone who already reports to them, or someone who has been
            deactivated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/users/{user_id}/reset-password:
    post:
      tags: [Users]
      summary: Issue a new password for a user who cannot sign in
      description: >
        The only route back into a locked-out account: change-password needs
        the password the user has lost, and no mail sender is configured.
        The new password is returned once and is never stored. The reset also
        drops every refresh token the user holds and rotates their calendar
        feed URL. A UNIT_ADMIN is limited to their own unit and may not reset
        an admin account.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A new password, shown this one time.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{user_id}/deactivate:
    post:
      tags: [Users]
      summary: Stop an account without deleting it
      description: >
        For somebody who has left. Every record against the account stays,
        and every way in closes: signing in, refreshing, an access token
        already issued, an API key bound to the account, the calendar feed
        and an open WebSocket. The account drops out of `GET /users/` unless
        `include_deactivated=true`, out of the staff directory and staff
        locations, and cannot be named as a manager, lead or substitute, sent
        a guest, issued a new API key or enrolled by a CSV import. Its refresh
        tokens are dropped and its calendar feed URL is rotated. API keys are
        held rather than deleted, and the email address stays taken.


        What already names the account is left alone. A slot it leads keeps
        producing days with it as the lead, and an absence request already
        sent to it stays pending, since only the manager a request went to can
        decide it; the requester submits it again once they have a new manager.


        Calling it on an account that is already deactivated changes nothing.
        A UNIT_ADMIN is limited to their own unit and may not deactivate an
        admin account, and nobody may deactivate their own.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The account, with `deactivated_at` set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: The caller named their own account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/users/{user_id}/reactivate:
    post:
      tags: [Users]
      summary: Let a deactivated account back in
      description: >
        The password and API keys work again as they were. The calendar feed
        URL does not: deactivating rotated it, so the person fetches the new
        one. If the old password should not work again, a reset-password
        afterwards issues a new one. Calling it on an active account changes
        nothing. A UNIT_ADMIN has the same limits as for deactivating.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The account, with `deactivated_at` null.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{user_id}/status:
    put:
      tags: [Users]
      summary: Update staff availability status
      description: Staff update their own. A super-admin can update anyone.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Status updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── API keys ─────────────────────────────
...